Moved IgniteDriver into Ignite module
fix #1520
