+++ simulator/src/main/java/com/hazelcast/simulator/vendors/VendorDriver.java
@@ -44,10 +44,21 @@ public abstract class VendorDriver<V> implements Closeable {
 
         if (vendorName.equals("hazelcast") || vendorName.equals("hazelcast-enterprise")) {
             return new HazelcastDriver();
-        } else if (vendorName.equals("ignite")) {
-            return new IgniteDriver();
         } else {
-            throw new IllegalArgumentException("Unknown vendor [" + vendorName + "]");
+            String driverName = "com.hazelcast.simulator." + vendorName + "."
+                    + vendorName.substring(0, 1).toUpperCase() + vendorName.substring(1) + "Driver";
+            Class driverClass;
+            try {
+                driverClass = VendorDriver.class.getClassLoader().loadClass(driverName);
+            } catch (ClassNotFoundException e) {
+                throw new CommandLineExitException(format("Could not locate driver class [%s]", driverName));
+            }
+
+            try {
+                return (VendorDriver) driverClass.newInstance();
+            } catch (Exception e) {
+                throw new CommandLineExitException(format("Failed to create an instance of driver [%s]", driverName), e);
+            }
         }
     }
